Rename layers to mappings
Odium is going to be composed of 4 things:

* Base options, similar to vim-sensible.
* Hand picked plugins.
* `config-` files for global plugin configurations.
* `language-` files for global per-language (usually plugin)
  configurations.
* `mappings-` files. This is new.

Mapping files are a layer between the user and the plugins. These files
will contain consistent, ergonomic, mnemonic mappings to perform
operations.

How each operation is implemented, should be abstracted from the user.

For example, within odium you can search files. The mapping for this is
`<spc> s f`. The fact that it's implemented using FZF is an
implementation detail.