Fix indentation for gitcommit and markdown
Both gitcommit and markdown are similar regarding the formatting that
the body allows.

In particular, this commit enables:

* `t`: text wrapping.
* `c`: inserting comment leader on text wrapping.
* `q`: formatting of comments.
* `j`: remove comment leader when joining lines.

But, it sets the possible comments only for quotes (`b:>`) and not for
lists. Therefore, list bullets won't be automatically added (nor
removed) to avoid confusion.

Also moves all ftplugin files to the after directory to make sure the
settings apply.